requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.1
tensorflow==2.13.0
scikit-image==0.21.0
matplotlib==3.7.2
werkzeug==2.3.7
gunicorn==21.2.0